+++ pallets/support/src/traits.rs
@@ -115,11 +115,11 @@ pub trait PricesInterface {
 	fn get_last_traded_price(market_id: u128) -> FixedI128;
 	fn update_last_traded_price(market_id: u128, price: FixedI128);
 	fn get_remaining_markets() -> Vec<u128>;
-	fn get_no_of_batches_for_current_epoch() -> u128;
+	fn get_no_of_batches_for_current_epoch() -> u64;
 	fn get_last_abr_timestamp() -> u64;
 	fn get_next_abr_timestamp() -> u64;
 	fn get_previous_abr_values(starting_epoch: u64, market_id: u128, n: u64) -> Vec<ABRDetails>;
-	fn get_remaining_pay_abr_calls() -> u128;
+	fn get_remaining_pay_abr_calls() -> u64;
 }
 
 pub trait FixedI128Ext {
